failed
Failed in the pause script, the script didn't work with the keycode Tab. But the button control menu replaces it work.

And monster box, actually working but it very stupid that it only stay on the birth place rather than move, and my tried to type the code to let them chasing player, failed at all!