failure
In Pause.cs, it wasn't work when click the button "Tab" after I coded the BUttonPause.cs. Before BUttonPause, it worked. May be the two scripts conflicted.

And monster box, the script only can create prefabs(clone) of monster, the monster were stupid that cannot move. The issue may happened in the axis of the birthplace.

Last one was the UI, in game UI was too small instead of can cover all the screen so that the pause button was too long to use, and the camera was following the mouse, so it's very dizzy(I forgot to set the speed of mouse, but my laptop was too trash that I can adapt the frame)